+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+## Added
+- Handling `Symfony\Component\Validator\Exception\ValidationFailedException` while resolving entryDto if you prefer to use constraints from `symfony/validation`. 
 
 ## [3.0.2] - 2021-11-08
 ## Changed